+++ packages/client/src/gen/video/sfu/event/events.ts
@@ -6,6 +6,7 @@ import {
   CallEndedReason,
   CallGrants,
   CallState,
+  ClientCapability,
   ClientDetails,
   Codec,
   ConnectionQuality,
@@ -254,6 +255,15 @@ export interface SfuEvent {
          * @generated from protobuf field: stream.video.sfu.event.ChangePublishOptions change_publish_options = 27;
          */
         changePublishOptions: ChangePublishOptions;
+      }
+    | {
+        oneofKind: 'inboundStateNotification';
+        /**
+         * InboundStateNotification
+         *
+         * @generated from protobuf field: stream.video.sfu.event.InboundStateNotification inbound_state_notification = 28;
+         */
+        inboundStateNotification: InboundStateNotification;
       }
     | {
         oneofKind: undefined;
@@ -508,6 +518,10 @@ export interface JoinRequest {
    * @generated from protobuf field: repeated stream.video.sfu.models.SubscribeOption preferred_subscribe_options = 10;
    */
   preferredSubscribeOptions: SubscribeOption[];
+  /**
+   * @generated from protobuf field: repeated stream.video.sfu.models.ClientCapability capabilities = 11;
+   */
+  capabilities: ClientCapability[];
 }
 /**
  * @generated from protobuf message stream.video.sfu.event.ReconnectDetails
@@ -875,6 +889,36 @@ export interface CallEnded {
    */
   reason: CallEndedReason;
 }
+/**
+ * @generated from protobuf message stream.video.sfu.event.InboundStateNotification
+ */
+export interface InboundStateNotification {
+  /**
+   * @generated from protobuf field: repeated stream.video.sfu.event.InboundVideoState inbound_video_states = 1;
+   */
+  inboundVideoStates: InboundVideoState[];
+}
+/**
+ * @generated from protobuf message stream.video.sfu.event.InboundVideoState
+ */
+export interface InboundVideoState {
+  /**
+   * @generated from protobuf field: string user_id = 1;
+   */
+  userId: string;
+  /**
+   * @generated from protobuf field: string session_id = 2;
+   */
+  sessionId: string;
+  /**
+   * @generated from protobuf field: stream.video.sfu.models.TrackType track_type = 3;
+   */
+  trackType: TrackType;
+  /**
+   * @generated from protobuf field: bool paused = 4;
+   */
+  paused: boolean;
+}
 // @generated message type with reflection information, may provide speed optimized methods
 class SfuEvent$Type extends MessageType<SfuEvent> {
   constructor() {
@@ -1032,6 +1076,13 @@ class SfuEvent$Type extends MessageType<SfuEvent> {
         kind: 'message',
         oneof: 'eventPayload',
         T: () => ChangePublishOptions,
+      },
+      {
+        no: 28,
+        name: 'inbound_state_notification',
+        kind: 'message',
+        oneof: 'eventPayload',
+        T: () => InboundStateNotification,
       },
     ]);
   }
@@ -1341,6 +1392,17 @@ class JoinRequest$Type extends MessageType<JoinRequest> {
         kind: 'message',
         repeat: 2 /*RepeatType.UNPACKED*/,
         T: () => SubscribeOption,
+      },
+      {
+        no: 11,
+        name: 'capabilities',
+        kind: 'enum',
+        repeat: 1 /*RepeatType.PACKED*/,
+        T: () => [
+          'stream.video.sfu.models.ClientCapability',
+          ClientCapability,
+          'CLIENT_CAPABILITY_',
+        ],
       },
     ]);
   }
@@ -1787,3 +1849,45 @@ class CallEnded$Type extends MessageType<CallEnded> {
  * @generated MessageType for protobuf message stream.video.sfu.event.CallEnded
  */
 export const CallEnded = new CallEnded$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class InboundStateNotification$Type extends MessageType<InboundStateNotification> {
+  constructor() {
+    super('stream.video.sfu.event.InboundStateNotification', [
+      {
+        no: 1,
+        name: 'inbound_video_states',
+        kind: 'message',
+        repeat: 2 /*RepeatType.UNPACKED*/,
+        T: () => InboundVideoState,
+      },
+    ]);
+  }
+}
+/**
+ * @generated MessageType for protobuf message stream.video.sfu.event.InboundStateNotification
+ */
+export const InboundStateNotification = new InboundStateNotification$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class InboundVideoState$Type extends MessageType<InboundVideoState> {
+  constructor() {
+    super('stream.video.sfu.event.InboundVideoState', [
+      { no: 1, name: 'user_id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
+      { no: 2, name: 'session_id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
+      {
+        no: 3,
+        name: 'track_type',
+        kind: 'enum',
+        T: () => [
+          'stream.video.sfu.models.TrackType',
+          TrackType,
+          'TRACK_TYPE_',
+        ],
+      },
+      { no: 4, name: 'paused', kind: 'scalar', T: 8 /*ScalarType.BOOL*/ },
+    ]);
+  }
+}
+/**
+ * @generated MessageType for protobuf message stream.video.sfu.event.InboundVideoState
+ */
+export const InboundVideoState = new InboundVideoState$Type();