+++ packages/react-native-sdk/src/components/Participant/ParticipantView/VideoRenderer.tsx
@@ -5,6 +5,7 @@ import { RTCView } from '@stream-io/react-native-webrtc';
 import type { ParticipantViewProps } from './ParticipantView';
 import {
   CallingState,
+  hasPausedTrack,
   hasScreenShare,
   hasVideo,
   SfuModels,
@@ -99,9 +100,10 @@ export const VideoRenderer = ({
     !!videoStreamToRender &&
     isVisible &&
     isPublishingVideoTrack &&
+    !hasPausedTrack(participant, trackType) &&
     isParticipantVideoEnabled(participant.sessionId);
 
-  React.useEffect(() => {
+  useEffect(() => {
     if (
       Platform.OS === 'ios' &&
       registerIosScreenshot &&